ensure site builds with standard plugins
Currently, javadoc is configured to run test-aggregate in this module, which
will fork a lifecycle that runs the unpack goal - however as it is not a full
package lifecycle, it will not have the dependencies it needs. This works around
the issue by placing both goals in prepare-package. Note that this means that
the result is not available to the site - if that's needed, the execution must
also be configured in the pre-site lifecycle.